==> course completion/modeling_and_evaluation/base_model/05_mcfadden.ipynb <==
This script computes the McFadden's adjusted R^2 for the training sample, based on the base model.

import numpy as np
import pandas as pd

def mcfadden(real_y, pred_y, k):
    L = np.where(real_y == 1, pred_y, 1 - pred_y)
    l = np.sum(np.log(L))
    L_null = np.where(real_y == 1, np.mean(real_y), 1 - np.mean(real_y))
    l_null = np.sum(np.log(L_null))
    return 1 - (l - k)/l_null

df = pd.read_stata("C:/Users/ys8mz/Downloads/all_predicted_scores_for_course_completion.dta")
df = df[df.original_sample == 'training']

k = 186

white = df[df.race == "white"]
print(mcfadden(np.array(white.binary_outcome), np.array(white.pred_score_1), k))
afam = df[df.race == "afam"]
print(mcfadden(np.array(afam.binary_outcome), np.array(afam.pred_score_1), k))
